+++ src/components/Warning.tsx
@@ -1,4 +1,3 @@
-// src/components/Warning.tsx
 import React from "react";
 import { FaExclamationTriangle } from "react-icons/fa";
 
@@ -11,7 +10,7 @@ const Warning: React.FC<WarningProps> = ({ children }) => {
     <div
       className="fixed top-20 left-1/2 transform -translate-x-1/2 z-50 
                     bg-yellow-400 border-l-4 border-red-500 
-                    text-red-500 px-20 py-5 rounded-md shadow-md"
+                    text-red-500 px-5 py-3 rounded-md shadow-md"
     >
       <div className="flex items-center">
         <FaExclamationTriangle className="mr-3 text-xl" />